requirements.txt
ib_insync>=0.9.86
APScheduler>=3.10.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas_market_calendars>=4.3.0
requests>=2.31.0
